abstraction of prefetch functions
